(By Sklower) checkpoint the current state of Cherenson's work.
SCCS-vsn: sys/netinet/in_pcb.c 7.24
SCCS-vsn: sys/netinet/ip_icmp.h 7.6
SCCS-vsn: sys/netinet/if_ether.h 7.9
SCCS-vsn: sys/netinet/ip_output.c 7.27
SCCS-vsn: sys/netinet/in.c 7.27
SCCS-vsn: sys/netinet/in_proto.c 7.9
SCCS-vsn: sys/netinet/tcp_usrreq.c 7.18
SCCS-vsn: sys/netinet/tcp_input.c 7.30
SCCS-vsn: sys/netinet/tcp_timer.c 7.20
SCCS-vsn: sys/netinet/ip_icmp.c 7.19
SCCS-vsn: sys/netinet/ip_var.h 7.9
SCCS-vsn: sys/netinet/tcp_subr.c 7.25
SCCS-vsn: sys/netinet/icmp_var.h 7.6
SCCS-vsn: sys/netinet/udp_usrreq.c 7.27
SCCS-vsn: sys/netinet/tcp.h 7.8
SCCS-vsn: sys/netinet/tcp_var.h 7.11
SCCS-vsn: sys/netinet/udp_var.h 7.8
SCCS-vsn: sys/netinet/igmp.c 7.3
SCCS-vsn: sys/netinet/ip_input.c 7.24
SCCS-vsn: sys/netinet/raw_ip.c 7.11
SCCS-vsn: sys/netinet/tcp_output.c 7.25
SCCS-vsn: sys/netinet/if_ether.c 7.27
SCCS-vsn: sys/netinet/igmp_var.h 7.2
SCCS-vsn: sys/netinet/ip_mroute.c 7.5

+++ usr/src/sys/netinet/icmp_var.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)icmp_var.h	7.5 (Berkeley) %G%
+ *	@(#)icmp_var.h	7.6 (Berkeley) %G%
  */
 
 /*
@@ -13,17 +13,17 @@
  */
 struct	icmpstat {
 /* statistics related to icmp packets generated */
-	int	icps_error;		/* # of calls to icmp_error */
-	int	icps_oldshort;		/* no error 'cuz old ip too short */
-	int	icps_oldicmp;		/* no error 'cuz old was icmp */
-	int	icps_outhist[ICMP_MAXTYPE + 1];
+	u_long	icps_error;		/* # of calls to icmp_error */
+	u_long	icps_oldshort;		/* no error 'cuz old ip too short */
+	u_long	icps_oldicmp;		/* no error 'cuz old was icmp */
+	u_long	icps_outhist[ICMP_MAXTYPE + 1];
 /* statistics related to input messages processed */
- 	int	icps_badcode;		/* icmp_code out of range */
-	int	icps_tooshort;		/* packet < ICMP_MINLEN */
-	int	icps_checksum;		/* bad checksum */
-	int	icps_badlen;		/* calculated bound mismatch */
-	int	icps_reflect;		/* number of responses */
-	int	icps_inhist[ICMP_MAXTYPE + 1];
+ 	u_long	icps_badcode;		/* icmp_code out of range */
+	u_long	icps_tooshort;		/* packet < ICMP_MINLEN */
+	u_long	icps_checksum;		/* bad checksum */
+	u_long	icps_badlen;		/* calculated bound mismatch */
+	u_long	icps_reflect;		/* number of responses */
+	u_long	icps_inhist[ICMP_MAXTYPE + 1];
 };
 
 #ifdef KERNEL
