+++ usr/src/sys/netinet/igmp_var.h
@@ -8,7 +8,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)igmp_var.h	7.1 (Berkeley) %G%
+ *	@(#)igmp_var.h	7.2 (Berkeley) %G%
  */
 
 /*
@@ -21,15 +21,15 @@
  */
 
 struct igmpstat {
-	u_int	igps_rcv_total;		/* total IGMP messages received */
-	u_int	igps_rcv_tooshort;	/* received with too few bytes */
-	u_int	igps_rcv_badsum;	/* received with bad checksum */
-	u_int	igps_rcv_queries;	/* received membership queries */
-	u_int	igps_rcv_badqueries;	/* received invalid queries */
-	u_int	igps_rcv_reports;	/* received membership reports */
-	u_int	igps_rcv_badreports;	/* received invalid reports */
-	u_int	igps_rcv_ourreports;	/* received reports for our groups */
-	u_int	igps_snd_reports;	/* sent membership reports */
+	u_long	igps_rcv_total;		/* total IGMP messages received */
+	u_long	igps_rcv_tooshort;	/* received with too few bytes */
+	u_long	igps_rcv_badsum;	/* received with bad checksum */
+	u_long	igps_rcv_queries;	/* received membership queries */
+	u_long	igps_rcv_badqueries;	/* received invalid queries */
+	u_long	igps_rcv_reports;	/* received membership reports */
+	u_long	igps_rcv_badreports;	/* received invalid reports */
+	u_long	igps_rcv_ourreports;	/* received reports for our groups */
+	u_long	igps_snd_reports;	/* sent membership reports */
 };
 
 #ifdef KERNEL
@@ -57,4 +57,3 @@ void	igmp_joingroup __P((struct in_multi *));
 void	igmp_leavegroup __P((struct in_multi *));
 void	igmp_fasttimo __P(());
 #endif
-
